Fix ThreadsDensitySetting rename conflict
