+++ app/src/main/java/com/infomaniak/mail/ui/main/settings/SettingsFragment.kt
@@ -30,7 +30,6 @@ import com.infomaniak.mail.data.cache.userInfos.UserPreferencesController
 import com.infomaniak.mail.data.models.UiSettings
 import com.infomaniak.mail.data.models.user.UserPreferences
 import com.infomaniak.mail.databinding.FragmentSettingsBinding
-import com.infomaniak.mail.utils.notYetImplemented
 import com.infomaniak.mail.utils.AccountUtils
 import com.infomaniak.mail.utils.notYetImplemented
 import kotlinx.coroutines.Dispatchers
@@ -88,7 +87,7 @@ class SettingsFragment : Fragment() {
         settingsAppLockSwitch.setOnClickListener { if (settingsAppLockSwitch.isChecked) notYetImplemented() }
 
         settingsThreadListDensity.setOnClickListener {
-            safeNavigate(SettingsFragmentDirections.actionSettingsToListDensitySetting())
+            safeNavigate(SettingsFragmentDirections.actionSettingsToThreadsDensitySetting())
         }
 
         settingsTheme.setOnClickListener {
