--- app/src/main/res/layout/fragment_list_density_setting.xml
@@ -1,110 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-  ~ Infomaniak kMail - Android
-  ~ Copyright (C) 2022 Infomaniak Network SA
-  ~
-  ~ This program is free software: you can redistribute it and/or modify
-  ~ it under the terms of the GNU General Public License as published by
-  ~ the Free Software Foundation, either version 3 of the License, or
-  ~ (at your option) any later version.
-  ~
-  ~ This program is distributed in the hope that it will be useful,
-  ~ but WITHOUT ANY WARRANTY; without even the implied warranty of
-  ~ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  ~ GNU General Public License for more details.
-  ~
-  ~ You should have received a copy of the GNU General Public License
-  ~ along with this program.  If not, see <http://www.gnu.org/licenses/>.
-  -->
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.main.settings.ListDensitySettingFragment"
-    tools:ignore="ContentDescription,UseCompoundDrawables">
-
-    <com.google.android.material.appbar.AppBarLayout
-        android:id="@+id/appBar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
-
-        <com.google.android.material.appbar.MaterialToolbar
-            android:id="@+id/toolbar"
-            android:layout_width="match_parent"
-            android:layout_height="?attr/actionBarSize">
-
-            <TextView
-                style="@style/H2"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginHorizontal="@dimen/marginStandardMedium"
-                android:ellipsize="end"
-                android:gravity="center_horizontal"
-                android:maxLines="1"
-                android:text="@string/settingsThreadListDensityTitle" />
-
-        </com.google.android.material.appbar.MaterialToolbar>
-
-    </com.google.android.material.appbar.AppBarLayout>
-
-    <androidx.core.widget.NestedScrollView
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:animateLayoutChanges="true"
-        android:scrollbars="none"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/appBar">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:orientation="vertical">
-
-            <TextView
-                style="@style/SettingHeader"
-                android:text="@string/settingsSelectDisplayModeDescription" />
-
-            <com.google.android.material.button.MaterialButtonToggleGroup
-                android:id="@+id/listDensityButtonsGroup"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:paddingHorizontal="@dimen/marginStandard"
-                android:paddingVertical="@dimen/marginStandardSmall"
-                app:checkedButton="@id/listDensityButtonNormal"
-                app:selectionRequired="true"
-                app:singleSelection="true">
-
-                <com.google.android.material.button.MaterialButton
-                    android:id="@+id/listDensityButtonCompact"
-                    style="@style/SettingButtonToggle"
-                    android:text="@string/settingsDensityOptionCompact" />
-
-                <com.google.android.material.button.MaterialButton
-                    android:id="@+id/listDensityButtonNormal"
-                    style="@style/SettingButtonToggle"
-                    android:text="@string/settingsDefault" />
-
-                <com.google.android.material.button.MaterialButton
-                    android:id="@+id/listDensityButtonLarge"
-                    style="@style/SettingButtonToggle"
-                    android:text="@string/settingsDensityOptionLarge" />
-            </com.google.android.material.button.MaterialButtonToggleGroup>
-
-            <ImageView
-                android:id="@+id/listDensityImage"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:paddingHorizontal="@dimen/marginStandard"
-                android:paddingVertical="@dimen/marginStandardSmall"
-                android:src="@drawable/bg_list_density_default" />
-
-        </LinearLayout>
-
-    </androidx.core.widget.NestedScrollView>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
